feat: Refactor and cleanup transactions handling code.
